Refactored front-end to add hierarchy between components
