+++ gwennosekai-frontend/src/app/app.module.ts
@@ -8,17 +8,17 @@ import { ROUTES } from './app.routes';
 
 
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './header/header.component';
-import { MenuComponent } from './menu/menu.component';
-import { PostListComponent } from './post-list/post-list.component';
-import { PostComponent } from './post/post.component';
-import { LogoComponent } from './logo/logo.component';
+import { HeaderComponent } from './common/header/header.component';
+import { MenuComponent } from './common/menu/menu.component';
+import { PostListComponent } from './front-office/post-list/post-list.component';
+import { PostComponent } from './front-office/post-list/post/post.component';
+import { LogoComponent } from './common/logo/logo.component';
 import {PostService} from "./post.service";
 import { FrontOfficeComponent } from './front-office/front-office.component';
 import { BackOfficeComponent } from './back-office/back-office.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { AdminMenuComponent } from './admin-menu/admin-menu.component';
-import { AdminArticleListComponent } from './admin-article-list/admin-article-list.component';
+import { DashboardComponent } from './back-office/dashboard/dashboard.component';
+import { AdminMenuComponent } from './back-office/admin-menu/admin-menu.component';
+import { AdminArticleListComponent } from './back-office/admin-article-list/admin-article-list.component';
 
 @NgModule({
   declarations: [
