+++ gwennosekai-frontend/src/app/app.routes.ts
@@ -1,8 +1,8 @@
 import {Routes} from '@angular/router';
 import {FrontOfficeComponent} from "./front-office/front-office.component";
 import {BackOfficeComponent} from "./back-office/back-office.component";
-import {DashboardComponent} from "./dashboard/dashboard.component";
-import {AdminArticleListComponent} from "./admin-article-list/admin-article-list.component";
+import {DashboardComponent} from "./back-office/dashboard/dashboard.component";
+import {AdminArticleListComponent} from "./back-office/admin-article-list/admin-article-list.component";
 
 export const ROUTES: Routes = [
   {
