+++ gwennosekai-frontend/src/app/back-office/dashboard/dashboard.component.ts
