+++ gwennosekai-frontend/src/app/common/header/header.component.scss
