+++ gwennosekai-frontend/src/app/common/logo/logo.component.spec.ts
