+++ gwennosekai-frontend/src/app/common/logo/logo.component.ts
