+++ gwennosekai-frontend/src/app/common/menu/menu.component.spec.ts
