+++ gwennosekai-frontend/src/app/front-office/post-list/post-list.component.html
