+++ gwennosekai-frontend/src/app/front-office/post-list/post/post.component.scss
